Hold an instant as a whole number of milliseconds
The calendars and the time zones held an instant as a double, which is what ICU does, and a
double stops being able to hold every millisecond about 285000 years from the epoch. The range
that the calendars cover reaches twenty times further than that, and the range of a TIMESTAMPTZ
reaches just past it, so the loss was reachable from SQL: the last microsecond of the range did
not survive a round trip, and consecutive milliseconds up there landed on the same value.

An instant is now an int64_t of milliseconds, which is exact over the whole range and still
reaches far beyond it. Microseconds are not the unit, because they would only reach a fifth of
the way through the range the calendars support, and because the calendars work in milliseconds
anyway - the layer above them already splits an instant into milliseconds and a remainder, and
was converting the exact count it had into a double and back.

Only the astronomy stays in floating point: the position of the sun and the moon are computed
from series expansions, and the instant of a new moon is genuinely not a whole millisecond. The
boundary is now the only place an instant is not an integer.

Some tests recorded the old behaviour, and two of them said so: one noted that a millisecond
below the minimum "succeeds because we are below the precision of UDate", and another that a
block of results was "off by 1ms" for the same reason. Those now give the answers a TIMESTAMP
gives for the same arithmetic, which is what they should have given all along.

Everything away from the limits is unchanged: the offsets of all 638 zones over sixty instants
each, and all 768000 date_diff values, are the same as before.

+++ extension/icu/icu-dateadd.cpp
@@ -87,7 +87,7 @@ timestamp_tz_t ICUCalendarAdd::Operation(timestamp_tz_t timestamp, interval_t in
 	Timestamp::Convert(timestamp_t(us), d, t);
 
 	// Now use the calendar to add the other parts
-	calendar->SetTime(double(millis));
+	calendar->SetTime(millis);
 
 	// Break units apart to avoid overflow
 	auto interval_h = interval.micros / Interval::MICROS_PER_MSEC;
