+++ extension/icu/icu-datefunc.cpp
@@ -77,7 +77,7 @@ void ICUDateFunc::SetTimeZone(Calendar *calendar, const string_t &tz_id, string 
 
 timestamp_tz_t ICUDateFunc::GetTimeUnsafe(Calendar *calendar, uint64_t micros) {
 	// Extract the new time
-	const auto millis = int64_t(calendar->GetTime());
+	const auto millis = calendar->GetTime();
 	if (calendar->HasFailed()) {
 		throw InternalException("Unable to get calendar time.");
 	}
@@ -86,12 +86,12 @@ timestamp_tz_t ICUDateFunc::GetTimeUnsafe(Calendar *calendar, uint64_t micros) {
 
 bool ICUDateFunc::TryGetTime(Calendar *calendar, uint64_t micros, timestamp_tz_t &result) {
 	// Extract the new time
-	auto millis = int64_t(calendar->GetTime());
+	auto millis = calendar->GetTime();
 	if (calendar->HasFailed()) {
 		return false;
 	}
 
-	// The time is a double, so it can't overflow (it just loses accuracy), but converting back to µs can.
+	// The time is a whole number of milliseconds, so converting it back to microseconds can overflow.
 	if (!TryMultiplyOperator::Operation<int64_t, int64_t, int64_t>(millis, Interval::MICROS_PER_MSEC, millis)) {
 		return false;
 	}
@@ -147,7 +147,7 @@ uint64_t ICUDateFunc::SetTime(Calendar *calendar, timestamp_tz_t date) {
 		micros += Interval::MICROS_PER_MSEC;
 	}
 
-	calendar->SetTime(double(millis));
+	calendar->SetTime(millis);
 	return uint64_t(micros);
 }
 
@@ -159,7 +159,7 @@ uint64_t ICUDateFunc::SetTimeNS(Calendar *calendar, timestamp_tz_ns_t date) {
 		nanos += Interval::NANOS_PER_MSEC;
 	}
 
-	calendar->SetTime(double(millis));
+	calendar->SetTime(millis);
 	return uint64_t(nanos);
 }
 
@@ -174,7 +174,7 @@ int32_t ICUDateFunc::ExtractField(Calendar *calendar, CalendarField field) {
 
 int32_t ICUDateFunc::SubtractField(Calendar *calendar, CalendarField field, timestamp_tz_t end_date) {
 	const int64_t millis = end_date.value / Interval::MICROS_PER_MSEC;
-	const auto sub = calendar->FieldDifference(double(millis), field);
+	const auto sub = calendar->FieldDifference(millis, field);
 	if (calendar->HasFailed()) {
 		throw ConversionException("Unable to subtract calendar part");
 	}
